Fixed truncating isuue with reading blocks with uint16_t

+++ audio_device.h
@@ -8,8 +8,8 @@
 
 extern HWAVEOUT hwaveout_device;
 void open_audio_dev(WAV_HEADER *info);
-void write_audioblock(HWAVEOUT audio_dev,LPSTR blocks,uint16_t size);
-LPSTR load_audio_block(char *filename,uint16_t *block_size);
+void write_audioblock(HWAVEOUT audio_dev,LPSTR blocks,size_t size);
+LPSTR load_audio_block(char *filename,size_t *block_size);
 char *waveout_error_codes(unsigned int Code);
 
 #endif
